fix object warning

+++ ros1_ws/src/omni3d_ros/src/DatasetUtils.py
@@ -666,7 +666,7 @@ def getComposedTransformOmni(origin, gt, pose, R):
     pose_ = pose
     pose_[3] = 1
     
-    pose_ = np.array([pose]).T
+    pose_ = np.array([pose], dtype=np.float32).T
     new_pose = T_inv @ pose_
 
     T = np.eye(4)
